feat: adds with login card hoc
